requirements.txt
Django==4.0.4
django-crontab==0.7.1
djangorestframework==3.13.1
requests==2.27.1
bs4==0.0.1